requirements.txt
flask>=3.0
twilio>=9.0
anthropic>=0.30
python-dotenv>=1.0
gunicorn>=21.0
